Create python.py
Write a Python program which accepts the radius of a circle from the user and computes area. Write a Python program to accept a filename from the user and print the extension of that. write a Python program to accept a filename from the user and print the extension of that.